+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Todo Tree Change Log
 
+## v0.0.171 - 2020-03-29
+
 - Refresh files after a timeout
 - Fix tag counts in tags only view
 - Add support for hiding tags from the status bar counts